+++ tests/testthat/test-gips_class.R
@@ -148,8 +148,37 @@ test_that("Properly validate the gips class with no optimization or after a sing
   expect_silent(validate_gips(g3))
   expect_silent(validate_gips(g_BF))
   expect_silent(validate_gips(g_BF_prob))
+  
+  
+  
+  # All NaN's
+  g_small <- gips(
+    S[1:4, 1:4], number_of_observations,
+    was_mean_estimated = FALSE
+  )
+  attr(g_small,"D_matrix") <- attr(g1,"D_matrix")[1:4,1:4] * 10000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000
+  
+  g_err <- g_small
+  expect_warning(expect_error(g_err <- find_MAP(
+    g_err, show_progress_bar = FALSE,
+    optimizer = "BF", return_probabilities = FALSE
+  ), "value of NaN occured"))
+  
+  g_err <- g_small
+  expect_warning(expect_error(g_err <- find_MAP(
+    g_err, max_iter = 2, show_progress_bar = FALSE,
+    optimizer = "MH", return_probabilities = FALSE
+  ), "value of NaN occured"))
+  
+  g_err <- g_small
+  expect_warning(expect_error(g_err <- find_MAP(
+    g_err, max_iter = 2, show_progress_bar = FALSE,
+    optimizer = "HC", return_probabilities = FALSE
+  ), "value of NaN occured"))
+  
 
 
+  # Other tests
   g_err <- g2
   class(g_err[[1]]) <- "test"
   expect_error(