+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Sometimes `post_probabilities` underflows to 0. This is appropriately validated now.
 * Typos in Vignettes
+* Proper testing of examples
 
 
 # gips 1.0.0