+++ NEWS.md
@@ -5,6 +5,7 @@
 * Sometimes `post_probabilities` underflows to 0. This is appropriately validated now.
 * Typos in Vignettes
 * Proper testing of examples
+* When `NaN`s occur in `find_MAP`, they will throw an error (used to show a warning).
 
 
 # gips 1.0.0