+++ NEWS.md
@@ -7,6 +7,7 @@
 * Proper testing of examples
 * When `NaN`s occur in `find_MAP`, they will throw an error (used to show a warning).
 * `NaN`s should not occur for `D_matrix <- diag(ncol(S)) * D_coef` when `1000 < D_coef < 1e300`.
+* `Inf` properly handled in `print.gips()`
 
 
 # gips 1.0.0